``` feat: add wavy overlap background for copus editor theme ```
