+++ src/index.tsx
@@ -33,8 +33,9 @@ import { ToolbarConfig } from './plugins/ToolbarPlugin';
 import { $createTextNodeX, TextNodeX } from './nodes/TextNodeX';
 import { MarkNodeX, MarkXType } from './nodes/MarkNodeX';
 import { $wrapSelectionInMarkNode, MarkNode } from '@lexical/mark';
+import { createUUID } from './utils/copus';
 
-function handleCopusSource(initialCopusSource: string) {
+function handleCopusSource(initialCopusSource: string, createMark?: (params: MarkXType) => Promise<MarkXType>) {
   return (editor: LexicalEditor) => {
     const root = $getRoot();
     initialCopusSource.split('\n').forEach((line) => {
@@ -43,7 +44,7 @@ function handleCopusSource(initialCopusSource: string) {
       root.append(paragraph);
     });
 
-    const allTextNodes = root.getAllTextNodes();
+    const allTextNodes = root.getAllTextNodes() as TextNodeX[];
     const anchor = allTextNodes[0];
     const focus = allTextNodes.length > 1 ? allTextNodes[allTextNodes.length - 1] : anchor;
     const rangeSelection = $createRangeSelection();
@@ -51,8 +52,21 @@ function handleCopusSource(initialCopusSource: string) {
     if (!$isRangeSelection(rangeSelection)) {
       return;
     }
-    $wrapSelectionInMarkNode(rangeSelection, false, '345', (ids) => {
-      return new MarkNodeX({ ids, source: true });
+
+    createMark?.({
+      opusUuid: createUUID(),
+      startNodeId: anchor.getId(),
+      startNodeAt: 0,
+      endNodeId: focus.getId(),
+      endNodeAt: focus.getTextContentSize(),
+    }).then((mark) => {
+      editor.update(() => {
+        if (mark?.id) {
+          $wrapSelectionInMarkNode(rangeSelection, false, mark.id, (ids) => {
+            return new MarkNodeX({ ids, source: true });
+          });
+        }
+      });
     });
   };
 }
@@ -61,16 +75,22 @@ export interface EditorProps {
   readOnly?: boolean;
   onChange?: (editorState: EditorState, html: string) => void;
   initialValue?: InitialConfigType['editorState'];
-  initialCopusSource?: string;
   toolbar?: ToolbarConfig;
   showLabel?: boolean;
-  markList?: MarkXType[];
-  copusCopy?: (params: MarkXType) => void;
+  copus?: {
+    markList?: MarkXType[];
+    copusCopy?: (params: MarkXType) => void;
+    initialSource?: string;
+    createMark?: (params: MarkXType) => Promise<MarkXType>;
+  };
 }
 
 function App(props: EditorProps): JSX.Element {
+  const { copus } = props;
+
   const initialConfig = {
-    editorState: props.initialValue ?? (props.initialCopusSource ? handleCopusSource(props.initialCopusSource) : null),
+    editorState:
+      props.initialValue ?? (copus?.initialSource ? handleCopusSource(copus.initialSource, copus.createMark) : null),
     namespace: 'CopusEditor',
     nodes: [
       ...PlaygroundNodes,
@@ -107,8 +127,7 @@ function App(props: EditorProps): JSX.Element {
                 readOnly={props.readOnly}
                 toolbar={props.toolbar}
                 showLabel={props.showLabel}
-                markList={props.markList}
-                copusCopy={props.copusCopy}
+                copus={props.copus}
               />
             </div>
           </SharedAutocompleteContext>
